LayoutTests/http/tests/loading/basic-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didFinishLoadForFrame
This is a test of load callbacks. It is only useful inside the regression test tool.
